backend/requirements.txt
flask
flask-cors
python-dotenv
transformers
torch
opencv-python
pillow
